+++ src/scripts/scrobble-timeline/2-merge-scrobbles.ts
@@ -1,32 +1,36 @@
-import {ScrobbleList} from 'src/types/scrobble';
+import {CompressedScrobbleList} from 'src/types/scrobble';
 
 import config from 'src/config';
 import {readAllJsonFiles, writeFile} from 'src/utils/file';
 import log, {proxyLogLength} from 'src/utils/log';
 
 import {loadCorrection} from 'src/ETL/extractors/correction';
+import {compress, uncompress} from 'src/ETL/transformers/compress';
 import {aggregatePlaycounts} from 'src/ETL/transformers/aggregate';
 import merge from 'src/ETL/transformers/merge-scrobbles';
 
-function extract(): Promise<ScrobbleList[]> {
-  return readAllJsonFiles<ScrobbleList>(config.scripts.scrobbleTimeline.fetchScrobbles.outputFilePath);
+function extract(): Promise<CompressedScrobbleList[]> {
+  return readAllJsonFiles<CompressedScrobbleList>(config.scripts.scrobbleTimeline.fetchScrobbles.outputFilePath);
 }
 
-function transform(scrobbleListList: ScrobbleList[]): Promise<ScrobbleList> {
+function transform(compressedScrobbleListList: CompressedScrobbleList[]): Promise<CompressedScrobbleList> {
   return loadCorrection(config.scripts.scrobbleTimeline.fetchScrobbles.corrections.artistName)
     .then((artistNameCorrection) => aggregatePlaycounts(
-      merge(scrobbleListList),
+      merge(compressedScrobbleListList.map(
+        (compressedScrobbleList) => compressedScrobbleList.map(uncompress),
+      )),
       artistNameCorrection,
-    ));
+    ))
+    .then((scrobbleList) => scrobbleList.map(compress));
 }
 
-function load(scrobbleList: ScrobbleList): Promise<ScrobbleList> {
+function load(compressedScrobbleList: CompressedScrobbleList): Promise<CompressedScrobbleList> {
   const {outputFilePath} = config.scripts.scrobbleTimeline.mergeScrobbles;
 
   log();
   log(`writing to "${outputFilePath}"`);
 
-  return writeFile(outputFilePath, scrobbleList);
+  return writeFile(outputFilePath, compressedScrobbleList);
 }
 
 extract()